:green_heart: Incorporate creation of config.json in Docker build process
The Docker build command now includes an argument to set the content of config.json. This file is directly generated within the Docker build process, removing the need to create it manually on the remote server during the deployment workflow. This change simplifies the deployment process and ensures all necessary files and settings for the app are included within the Docker image.